+++ src/test/services/DataTranformationServiceSuite.test.ts
@@ -10,23 +10,48 @@ describe("DataTransformationServiceSuite - Algebra", () => {
         const transformedData = await dtsService.timeUsedPerBoss(
             testGuildRaidData
         );
-        expect(transformedData).toEqual({
-            "TervigonLeviathan-Common": {
-                time: "00h 00m 00s",
-                tokens: 1,
+
+        expect(transformedData).toEqual([
+            {
+                "L0 Belisarius": {
+                    time: "00h 45m 27s",
+                    tokens: 4,
+                    bombs: 4,
+                    sideboss: [true, "L0 Belisarius"],
+                },
             },
-            "HiveTyrantLeviathan-Common": {
-                time: "00h 01m 01s",
-                tokens: 1,
+            "00h 45m 27s",
+        ]);
+    });
+
+    test("timeUsedPerBoss - should display primes as well if the option is enabled", async () => {
+        const transformedData = await dtsService.timeUsedPerBoss(
+            testGuildRaidData,
+            true
+        );
+
+        expect(transformedData).toEqual([
+            {
+                "L0 AdmecManipulus": {
+                    time: "00h 15m 07s",
+                    tokens: 1,
+                    bombs: 3,
+                    sideboss: [true, "L0 Belisarius"],
+                },
+                "L0 AdmecMarshall": {
+                    time: "00h 30m 03s",
+                    tokens: 2,
+                    bombs: 1,
+                    sideboss: [true, "L0 Belisarius"],
+                },
+                "L0 Belisarius": {
+                    time: "00h 00m 17s",
+                    tokens: 1,
+                    bombs: 0,
+                    sideboss: [false, "L0 Belisarius"],
+                },
             },
-            "HiveTyrantGorgon-Common": {
-                time: "00h 00m 59s",
-                tokens: 1,
-            },
-            "ScreamerKiller-Common": {
-                time: "00h 27m 32s",
-                tokens: 2,
-            },
-        });
+            "00h 45m 27s",
+        ]);
     });
 });